% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_county}
\alias{clean_county}
\title{Clean county population estimates data}
\usage{
clean_county(data, year)
}
\arguments{
\item{data}{Data frame read with read_county}

\item{year}{Vintage year}
}
\value{
Data frame
}
\description{
Transforms the data read with read_county so that each column is one
variable. Only returns selected race and hispanic origin estimates.
}
