% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_metro}
\alias{cw_metro}
\title{County-metro crosswalk}
\format{A data frame with four variables:
\describe{
\item{\code{county_fips}}{County FIPS code.}
\item{\code{county_name}}{County name.}
\item{\code{cbsa_code}}{Metro code.}
\item{\code{cbsa_name}}{Metro name.}
}}
\usage{
cw_metro
}
\description{
County-metro crosswalk
}
\keyword{datasets}
