% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popest-package.R
\docType{package}
\name{popest-package}
\alias{popest}
\alias{popest-package}
\title{popest: Tools to Get Population Estimates Data}
\description{
Tools to get US Census Bureau Population Estimates
    data.
}
\author{
\strong{Maintainer}: Ray Wright \email{raywright88@gmail.com}

}
\keyword{internal}
