% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_county}
\alias{read_county}
\title{Read county population estimates data}
\usage{
read_county(year)
}
\arguments{
\item{year}{Vintage year}
}
\value{
Data frame
}
\description{
Read county population estimates data
}
